Extended IR visitor framework (#24)
This extends the visitor framework to provide more control over the traversal, and to add hooks for visiting metadata nodes. It also introduces a visitor for traversing a pair of nodes in tandem.

Pre- and post-order traversal has been generalized: the framework now exposes `enter` and `leave` methods for each node type. Preorder traversal can be obtained by implementing `enter` methods, whereas postorder traversal is obtained by implementing `leave` methods. The visiting of each node type's children can also be customized.

Visit methods now return an enum value rather than a `bool` to indicate how traversal should proceed. Possibilities are `Continue` to continue with normal traversal; `SkipChildren` allows the `enter` method to specify that the children of the current node should not be visited; `Stop` to stop the traversal, but execute the `leave` methods on the way back up to the root; and `Abort` halts the traversal immediately.

Added `enter`, `leave`, and `visitChildren` methods for metadata IR nodes.