Add explicit go-pg type annotations for ServiceID (#103)
This is the result of adding a bulk update against a table with a ServiceID column. Without this type annotation on that table, the ServiceID in the bulk update is encoded as JSON, resulting in Postgres rejecting the update due to a type error. Furthermore, once this annotation is added to that table, it needs to be propagated to all ServiceID columns, or else joins don't work. 😞